Move wrapOk into cosmwasm module
